move request file format into own package
