+++ requestfiles/output_file.go
@@ -1,4 +1,4 @@
-package main
+package requestfiles
 
 import (
 	"encoding/gob"
